+++ src/builder/MD3.cc
@@ -26,7 +26,6 @@
 #include "builder/MD3.hh"
 
 #include "client/Context.hh"
-#include "client/OpenGL.hh"
 
 #include "builder/Compiler.hh"